JSSE14Factory renamed to JSSEFactory and the interface deleted
- We don't have any other implementation
